libstb/secvar: add secvar api implementation
This patch provides the OPAL runtime service frontend for the host OS to
retrieve secure variables, and append new ones for processing on the
next reboot. These calls operate on the internal abstraction or utilize
the platform-provided driver hooks, and therefore this API should not
need to be updated to support changes in storage or backend drivers.

Included are the following functions:
 - opal_secvar_get()
 - opal_secvar_get_next()
 - opal_secvar_enqueue_update()

opal_secvar_get() retrieves the data blob associated with a given key.
The data buffer may be set to NULL to only query for variable size. This
runtime service only operates on the variable bank.

opal_secvar_get_next() can be used to iterate through the list of
variable keys in the variable bank. Supplying an empty key (or zero key
length) returns the key of the first variable in the variable bank.
Supplying a valid key returns the key of the next variable in sequence.

opal_secvar_enqueue_update() provides a method for the host OS to submit
a new variable for processing on next boot, by appending it to the
update bank. As this does not affect the variable bank, appending a
variable via this runtime service will not affect the output of the
previous set of functions. The update queue is only processed during
secvar initialization.

V2:
 - removed opal_secvar_backend, replaced by DT node
 - removed unnecessary argument casting
 - all calls return OPAL_RESOURCE if secvar failed to init

V3:
 - remove metadata from API parameters
 - remove opal_secvar_get_size
 - change enqueue to replace an update with a repeat name, rather
    than enqueueing the duplicate
 - change enqueue to unstage an update matching a key if size is zero
 - make all key parameters const where possible
 - rename key_size to key_buf_size in _get_next
 - fix leaking node when enqueue could not allocate the secvar

V4:
 - enqueue update now uses secvar alloc/realloc
 - use storage-defined max var size instead of hardcoded constant

Signed-off-by: Eric Richter <erichte@linux.ibm.com>