witherspoon: enable secvar for witherspoon platform
Secure variable support needs to be enabled for each platform, and each
platform needs to select which storage and backend drivers to use (or
alternatively implement their own). This patch adds secure variable
support to the witherspoon platform.

NOTE: This patch includes commented out code to enable "Fake NV" mode,
intended for review purposes only. To review or test secure variables
on a non-witherspoon platform, replace this patch with a similar
one for your given platform with the Fake NV lines uncommented.

Signed-off-by: Eric Richter <erichte@linux.ibm.com>