secvar/storage: add draft secvar storage driver for pnor-based p9 platforms
This patch implements the platform specific logic for persisting the
secure variable storage banks across reboots via the SECBOOT PNOR
partition.

For POWER 9, all secure variables and updates are stored in the
in the SECBOOT PNOR partition. The partition is split into three
sections: two variable bank sections, and a section for storing
updates. The driver alternates writes between the two variable
sections, so that the final switch from one set of variables to
the next can be as atomic as possible by flipping an "active bit"
stored in TPM NV.

PNOR space provides no lock protection, so prior to writing the
variable bank, a sha256 hash is calculated and stored in TPM NV.
This hash is checked against when loading the variables from
PNOR to ensure consistency -- otherwise a failure is reported,
no keys are loaded (which should cause skiroot to refuse to
boot if secure boot support is enabled).

V2:
 - properly set DYN_ALLOC flag after loading from pnor
 - fixed "out of space" condition check in pnor writes

V4:
 - renamed to secboot_tpm
 - implement base tpm storage logic
 - calculates hash on writes, verifies hash on reads
 - alternate writing to two banks
 - use alloc_secvar instead of malloc
 - determines max variable size
 - add basic unit tests for exposed driver functions

Signed-off-by: Eric Richter <erichte@linux.ibm.com>