secvar/backend: add edk2 derived key updates processing
As part of secureboot key management, the scheme for handling key updates
is derived from tianocore reference implementation[1]. The wrappers for
holding the signed update is the Authentication Header and for holding
the public key certificate is ESL (EFI Signature List), both derived from
tianocore reference implementation[1].

This patch adds the support to process update queue. This involves:
1. Verification of the update signature using the key authorized as per the
key hierarchy
2. Handling addition/deletion of the keys
3. Support for dbx(blacklisting of hashes)
4. Validation checks for the updates
5. Supporting multiple ESLs for single variable both for update/verification
6. Timestamp check
7. Allowing only single PK
8. Failure Handling

[1] https://github.com/tianocore/edk2-staging.git

SQUASH: edk2: return early if the update bank list is empty

edk2: fixes edk2 to use local memory copy before writing to global variable bank

The process keeps updating the global copy of variable bank.
If the process exists in between, it would have over-written
the ones which are processed in the variable bank. It has two problems:

1.) We intend to succeed either for all or none
2.) Since for any process failure, the variable bank is not updated in PNOR,
it shows the updates for current boot, but they are lost in the next boot.

This patch fixes the problem by using an internal staging copy of variable
bank, and updating the global copy only if all the updates are successful.

V3:
 - performs operations on a copy, reverts to previous list on failure
 - refactor to pass banks as arguments to all process functions
 - physical presence support is handled by storage driver (CONFIRM, MAY
NOT HAVE BEEN IN V2)
 - increased error log verbosity

Signed-off-by: Nayna Jain <nayna@linux.ibm.com>
Signed-off-by: Eric Richter <erichte@linux.ibm.com>