secvar/backend: add edk2 derived key updates processing
As part of secureboot key management, the scheme for key updates
processing is derived from tianocore reference implementation[1].
This includes the verification of key updates signed in the form of
PKCS7 structure.

This patch adds the PKCS7 verification support for the signed
updates processed by the user. It also adds the preprocessing code
which initializes the empty non-volatile variables.

This patch is still a work-in-progress, for example. it still needs
to add the support for post-processing steps and better failure handling.

V2:
 - fixed memcpy based on sizeof(keylen) rather than the value
 - added version and compatible values to driver struct
 - renamed to edk2-compat

V4:
 - added draft document for general concepts
 - stores PK in the tpm
 - empty variables don't attempt to preallocate space
 - many malloc->zalloc changes
 - removed hard coded example variables
 - removed most printfs, converted some to prlog
 - updated for new pkcs7 implementation
 - removed unnecessary "SecureBoot", and "SecureMode"
     variables, handled by devtree entry

[1] https://github.com/tianocore/edk2-staging.git

Signed-off-by: Nayna Jain <nayna@linux.ibm.com>