secvar_tpmnv: add high-level tpm nv index abstraction for secvar
Multiple components, like the storage driver or backend driver, need to
store information in the one reserved TPM NV index. This abstraction
provides a method for components to share the index space without
stomping on each other's data, and without them needing to understand
anything about the other.

This is probably an overengineered solution to the problem, but the
intent is to keep the drivers as isolated from one another as possible.

Signed-off-by: Eric Richter <erichte@linux.ibm.com>