secvar_tpmnv: add high-level tpm nv index abstraction for secvar
Multiple components, like the storage driver or backend driver, may need
to store information in the one TPM NV index assigned for secure boot.
This abstraction provides a method for these components to share the
index space without stomping on each other's data, and without them
needing to understand anything about the other.

This is probably an overengineered solution to the problem, but the
intent is to keep the drivers as independent from one another as possible.

NOTE: At the moment, this implementation does NOT actually utilize the
TPM NV space. This patch depends on a forthcoming TSS integration to
handle the TPM communication. For now, this patch uses extra space at
the end of the SECBOOT PNOR partition as its simulated TPM NVRAM.
Temporary code is marked in the comments accordingly.

Signed-off-by: Eric Richter <erichte@linux.ibm.com>