secvar_tpmnv: add high-level tpm nv index abstraction for secvar
Multiple components, like the storage driver or backend driver, may need
to store information in the one TPM NV index assigned for secure boot.
This abstraction provides a method for these components to share the
index space without stomping on each other's data, and without them
needing to understand anything about the other.

This is probably an overengineered solution to the problem, but the
intent is to keep the drivers as independent from one another as possible.

NOTE: This version of the patch now includes the ability to switch between
a "Fake TPMNV" mode simulated using PNOR, and use of an actual TPM. The
simulated mode exists for unit test and review purposes on machines without
a TPM. It is not intended for production use.

Signed-off-by: Eric Richter <erichte@linux.ibm.com>