secvar_devtree: add generic compatible, and new format device tree property
This patch adds a generic compatible entry to the secvar/compatible
device tree property for kernels, etc to match for basic secvar
information (e.g. the linux secvar-sysfs implementation).

The new "format" property exposes the expected format for data passing
in and out of the secvar API. In most cases, this should be the same as
the backend-specific compatible entry.

Signed-off-by: Eric Richter <erichte@linux.ibm.com>