secvar/edk2: add support for physical presence
In case of loss of the private key part of PK, the administrator cannot
submit a signed update to replace his old PK. In such cases, he needs
a recovery mechanism to delete the old PK and submit a new one.

This patch adds the support for physical presence. The administrator
can now delete his PK by asserting the physical presence. This will
also delete all the lower hierarchy keys. The deletion of PK switches
the system to the setup mode. The administrator can now add his new PK
and all the subsequent keys.

Signed-off-by: Nayna Jain <nayna@linux.ibm.com>
Signed-off-by: Eric Richter <erichte@linux.ibm.com>