libstb/tss2: add skiboot wrappers for TSS commands
While the TSS does most of the heavy lifting for serialization and
deserialization of TSS commands, there is still a little bit of legwork to
utilize the library effectively.

This patch introduces a set of functions that simplify TPM commands to a
single function call that does not depend on any other TPM or TSS-related
headers.

Signed-off-by: Eric Richter <erichte@linux.ibm.com>