secvar_main: increase error verbosity, restyle all comments
This commit converts all the "//" style comments to the appropriate
C-style, adds a few comments, and also adds a couple error log outputs.

The following commits performs a minor refactor of the secvar_main
flow. This commit has been separated out to reduce the amount of noise
in those commits, and can likely be squashed if necessary.

V4:
 - adjusted lines to 80 columns

Signed-off-by: Eric Richter <erichte@linux.ibm.com>