secvar_main: rework secvar_main error flow, make storage locking explicit
This patch adjusts the behavior of secvar_main to actually halt the boot
in some form if there is an issue initializing secure variables. The secvar
storage driver contains the secure boot state, and therefore if that fails
to initialize, we immediately need to halt the boot. For all other cases we
enforce secure boot in the bootloader by setting the secure mode flag, but
booting with an empty keyring (and thus, cannot verify a kexec image).

Previously, the storage driver was expected to handle any locking
procedures implicitly as part of the write operation. This patch uses the
new lockdown hook which makes locking explicit and part of the secvar_main
flow.

The storage driver is now locked unconditionally when exiting
secvar_main, and the lockdown() call should halt the boot if it encounters
any sign of struggle.

V4:
 - adjusted lines to 80 columns
 - calls secureboot_enforce() instead of abort()
V5:
 - adjusted and expanded the comments to hopefully make the flow
    a little easier to follow
 - renamed .lock() to .lockdown()
 - renamed out to soft_fail

Signed-off-by: Eric Richter <erichte@linux.ibm.com>