edk2 file is refactored
The file is refactored into :
edk2-compat.c - just implements the backend hooks. And process hook now is calling
process_update() to process each update.
edk2-compat-util.c - implements all the utility functions for edk2 processing.
This is the file which can be used for pre-validation.
backend-common.c - This is the file which might have common functions across
backend. For eg.. I will be adding physical presence and hardware key hash
changes in the next patch.

key_equals() and checks in timestamp() can be refined further to use macros
but is not done here.

Signed-off-by: Nayna Jain <nayna@linux.vnet.ibm.com>