Fix the "possibly a problem" reported by BEAM.
It wasn't as such a problem in our case as we know the hierarchy
and the supported variables. Still, it was good to follow the
steps which keeps the code right in any case.

-- ERROR7     /*accessing beyond memory*/     >>>ERROR7_get_key_authority_df92eaf1505
"libstb/secvar/backend/edk2-compat-process.c", line 83: accessing beyond array bounds
ONE POSSIBLE PATH LEADING TO THE ERROR:
 "libstb/secvar/backend/edk2-compat-process.c", line 77: the if-condition is true
 "libstb/secvar/backend/edk2-compat-process.c", line 79: the if-condition is true
 "libstb/secvar/backend/edk2-compat-process.c", line 81: the if-condition is true
 "libstb/secvar/backend/edk2-compat-process.c", line 83: assigning into `i'
 "libstb/secvar/backend/edk2-compat-process.c", line 83: the index `i++' is outside the bounds [0,2] of array `ret'.

 The index has value = 3.

 VALUES AT THE END OF THE PATH:
  i = 4

Signed-off-by: Nayna Jain <nayna@linux.vnet.ibm.com>