edk2: fixes edk2 to use local memory copy before writing to global variable bank
The process keeps updating the global copy of variable bank.
If the process exists in between, it would have over-written
the ones which are processed in the variable bank. It has two problems:

1.) We intend to succeed either for all or none
2.) Since for any process failure, the variable bank is not updated in PNOR,
it shows the updates for current boot, but they are lost in the next boot.

This patch fixes the problem by using an internal staging copy of variable
bank, and updating the global copy only if all the updates are successful.

Signed-off-by: Nayna Jain <nayna@linux.ibm.com>