secvar/storage: add secvar storage driver for pnor-based p9
This patch implements the platform specific logic for persisting the
secure variable storage banks across reboots via the SECBOOT PNOR
partition.

For POWER 9, all secure variables and updates are stored in the
in the SECBOOT PNOR partition. The partition is split into three
sections: two variable bank sections, and a section for storing
updates. The driver alternates writes between the two variable
sections, so that the final switch from one set of variables to
the next can be as atomic as possible by flipping an "active bit"
stored in TPM NV.

PNOR space provides no lock protection, so prior to writing the
variable bank, a sha256 hash is calculated and stored in TPM NV.
This hash is compared against the hash of the variables loaded from
PNOR to ensure consistency -- otherwise a failure is reported, no keys
are loaded (which should cause skiroot to refuse to boot if secure boot
support is enabled).

Signed-off-by: Eric Richter <erichte@linux.ibm.com>

V3:
 - introduced priority variables (which are stored in TPM NV)
 - checks if nv indices are defined rather than relying on an error code
 - now handles physical presence, redefines NV indices if asserted
    (which forces full reformat, resets secure state)
 - checks if NV indices are defined with the correct attributes
 - added .lock() hook
 - increased error verbosity
 - addressed comments and reviews from the V2 set

V4:
 - add documentation for this driver
 - removed unused SECURE_STORAGE flag, renamed PRIORITY to PROTECTED
     This is still open to suggestions, PROTECTED sounded more clear)
 - fixed missing pointer advancement when writing protected variables
 - moved the expected NV public name hashes to seperate header
 - cleaned up most line lengths to under 80 columns, with the exception
     of prlog statements and possibly a few others.

V5:
 - overhauled serialization/deserialization:
   - no longer memcpy's full structs
   - always uses the tighter packing scheme
   - cleaned up/merged helper functions
   - writes all numeric values as BE
 - change some programming errors to asserts
 - moved the public name hashes back into the .c file