secvar: change backend hook interface to take in bank references
Previously, backends were implicitly expected to operate on global
references to the variable and update banks. This patch changes the
interface for this driver to instead take the banks in as an argument.

This removes the implict dependency on these references, makes the
design consistent with the storage driver, and also will simplify unit
testing of these functions.

V4:
 - squashed in a whitespace fix for the storage driver struct

V5:
 - adjusted comments so they don't blow way past the column limit

Signed-off-by: Nayna Jain <nayna@linux.ibm.com>
Signed-off-by: Eric Richter <erichte@linux.ibm.com>