secvar/backend: fix esl validation to allow
hash for dbx variable

Currently, validate_esl_list() function checks only for ESL data
as certificate. This results in invalidation of dbx data because
it is not cert type but is hash.

This patch fixes the issue by adding validation check for dbx.
It support SHA1/SHA224/SHA256/SHA384/SHA512 hashes for dbx.

Signed-off-by: Nayna Jain <nayna@linux.ibm.com>