backed/edk2: Reset the setup_mode in case of failure
The current code updates the setup_mode as soon as PK update
is processed. This is needed to ensure that sequential updates in
queue are processed as per new update. However, if any next update
fails, the setup_mode state is not reset to original. This patch
fixes to always set the setup_mode as per final version of PK in
variable_bank before exiting the process().

Signed-off-by: Nayna Jain <nayna@linux.ibm.com>