secvar/backend: Bugfixes regarding hw-key-hash in edk2 driver
HW-KEY-HASH defines the ownership of firmware and mapping of PK with it
associates the PK authority with the firmware owner. This patch fixes
following bugs in handling the mapping of PK with HW-KEY-HASH.
Additionally, it improves logs.

* Failure in adding/deleting PK as part of failure of processing any
subsequential update in the queue didn't reset the global variable
setup_mode to the original value. This patch adds the fix to always
set the value of setup_mode as per final contents in variable_bank
before existing process()

* Deletion of HWKH as part of deleting PK was only updating the value of
the variable to be zero. It didn't unallocate it. However, this didn't
deallocate the variable from the bank and was getting exposed via sysfs.

* The mismatch in verification of hw-key-hash, was also clearing staging
bank, which isn't initialized in this case. Fix the cleanup to only clear
update_bank.

* Fixes a memory leak in validate_esl_list().

Signed-off-by: Nayna Jain <nayna@linux.ibm.com>